Move Mongoid models into their own folder
