+++ lib/mongoid_models/client.rb
